major refactor of worker for build content
