+++ .gitignore
@@ -23,5 +23,5 @@ poetry.lock
 *_parsetab.py
 condaenv.n_dn60fa.requirements.txt
 worker/.worker_STABLE
-!worker/build
+!worker/datagen_build
 .DS_Store
