+++ worker/service/db_connector.py
@@ -1,30 +1,11 @@
-import os
-import uuid
 from abc import abstractmethod, ABC
-from asyncio import sleep
-from dataclasses import dataclass, asdict
 from datetime import datetime
-from enum import Enum
 from typing import *
 
-from dotenv import load_dotenv
 from pymongo import MongoClient
 from pymongo.collection import Collection
 from pymongo.database import Database
-# from biosimulators_processes import CORE
 
-
-# -- globally-shared content-- #
-
-
-load_dotenv('../assets/dev/config/.env_dev')
-
-DB_TYPE = "mongo"  # ie: postgres, etc
-DB_NAME = "service_requests"
-BUCKET_NAME = os.getenv("BUCKET_NAME")
-
-
-# -- shared functions -- #
 
 def check_jobs(coll):
     from main import db_connector as conn
@@ -42,62 +23,6 @@ def check_jobs(coll):
 
     return not_complete
 
-
-def unique_id():
-    return str(uuid.uuid4())
-
-
-def handle_exception(context: str) -> str:
-    import traceback
-    from pprint import pformat
-    tb_str = traceback.format_exc()
-    error_message = pformat(f"{context} error:\n{tb_str}")
-    
-    return error_message
-
-
-async def load_arrows(timer):
-    check_timer = timer
-    ell = ""
-    bars = ""
-    msg = "|"
-    n_ellipses = timer
-    log_interval = check_timer / n_ellipses
-    for n in range(n_ellipses):
-        single_interval = log_interval / 3
-        await sleep(single_interval)
-        bars += "="
-        disp = bars + ">"
-        if n == n_ellipses - 1:
-            disp += "|"
-        print(disp)
-
-
-# -- base python dataclass with to_dict() method -- #
-
-@dataclass
-class BaseClass:
-    """Base Python Dataclass multipurpose class with custom app configuration."""
-    def to_dict(self):
-        return asdict(self)
-
-
-# -- jobs -- #
-
-class JobStatus(Enum):
-    PENDING = "PENDING"
-    IN_PROGRESS = "IN_PROGRESS"
-    COMPLETED = "COMPLETED"
-    FAILED = "FAILED"
-
-
-class DatabaseCollections(Enum):
-    PENDING_JOBS = "PENDING_JOBS".lower()
-    IN_PROGRESS_JOBS = "IN_PROGRESS_JOBS".lower()
-    COMPLETED_JOBS = "COMPLETED_JOBS".lower()
-
-
-# -- database connectors: currently exclusive to mongodb. TODO: create a dbconnector for a relational db -- #
 
 class DatabaseConnector(ABC):
     """Abstract class that is both serializable and interacts with the database (of any type). """
